Rename zora_premint_v2 -> zora_premint
