+++ src/premints/mod.rs
@@ -1,2 +1,2 @@
 pub mod metadata;
-pub mod zora_premint_v2;
+pub mod zora_premint;
