+++ src/premints/zora_premint/rules.rs
@@ -1,5 +1,5 @@
 use crate::chain::view_contract_call;
-use crate::premints::zora_premint_v2::types::{
+use crate::premints::zora_premint::types::{
     IZoraPremintV2, ZoraPremintV2, PREMINT_FACTORY_ADDR,
 };
 use crate::rules::Evaluation::Accept;
