+++ tests/e2e_test.rs
@@ -16,8 +16,8 @@ use alloy::sol_types::{SolCall, SolValue};
 use alloy::transports::{RpcError, TransportErrorKind};
 use mintpool::config::{BootNodes, ChainInclusionMode, Config};
 use mintpool::controller::{ControllerCommands, DBQuery};
-use mintpool::premints::zora_premint_v2::types::IZoraPremintV2::MintArguments;
-use mintpool::premints::zora_premint_v2::types::{
+use mintpool::premints::zora_premint::types::IZoraPremintV2::MintArguments;
+use mintpool::premints::zora_premint::types::{
     IZoraPremintV2, ZoraPremintV2, PREMINT_FACTORY_ADDR,
 };
 use mintpool::rules::RulesEngine;
